refactor(service): split BpmnModelerService into focused adapters
Extract element-templates, settings, panel state, and clipboard handling
into dedicated service classes. This decouples session management from
cross-cutting concerns and lets each service own its configuration
subscriptions.

- BpmnElementTemplatesService: loads and broadcasts element templates
- BpmnSettingsBroadcaster: manages modeler settings and language
- BpmnPropertiesPanelService: owns panel visibility persistence
- BpmnClipboardMediator: mediates webview↔system clipboard access
- BpmnMigrationService: extracted workspace-wide diagram migration

Update BpmnEditorController to dispatch webview messages to the focused
services instead of routing everything through BpmnModelerService. Rename
subscribeToSettingChangeEvent to subscribeToConfigFolderChangeEvent since
other settings are now owned by BpmnSettingsBroadcaster.
