+++ .vscode/extensions.json
@@ -1,8 +1,3 @@
 {
-  "recommendations": [
-    "nrwl.angular-console",
-    "esbenp.prettier-vscode",
-    "vitest.explorer",
-    "dbaeumer.vscode-eslint"
-  ]
+    "recommendations": ["esbenp.prettier-vscode", "vitest.explorer", "dbaeumer.vscode-eslint"]
 }
