+++ apps/modeler-plugin/src/service/BpmnModelerService.ts
@@ -1,32 +1,14 @@
-import { posix } from "path";
-
-import {
-    BpmnFileQuery,
-    BpmnModelerSettingQuery,
-    ClipboardQuery,
-    ElementTemplatesQuery,
-    Engine,
-    LanguageQuery,
-    PropertiesPanelStateQuery,
-    TextClipboardQuery,
-} from "@miragon/bpmn-modeler-shared";
+import { BpmnFileQuery } from "@miragon/bpmn-modeler-shared";
 
 import { ModelerSession } from "../domain/session";
-import { SettingBuilder } from "../domain/model";
 import { ExecutionPlatformNotDetectedError, UserCancelledError } from "../domain/errors";
 import { getLatestVersion, getVersions } from "../domain/engineVersions";
 import { BpmnDocument } from "../domain/BpmnDocument";
-import { BpmnFileEntry, MigrationPlan, MigrationScope } from "../domain/MigrationPlan";
 import { EditorStore } from "../infrastructure/EditorStore";
-import { PropertiesPanelStateRepository } from "../infrastructure/PropertiesPanelStateRepository";
 import { VsCodeDocument } from "../infrastructure/VsCodeDocument";
-import { VsCodeWorkspace } from "../infrastructure/VsCodeWorkspace";
-import { VsCodeSettings } from "../infrastructure/VsCodeSettings";
 import { VsCodeStatusBar } from "../infrastructure/VsCodeStatusBar";
-import { VsCodeClipboard } from "../infrastructure/VsCodeClipboard";
 import { VsCodeNotifier } from "../infrastructure/VsCodeNotifier";
 import { VsCodePicker } from "../infrastructure/VsCodePicker";
-import { ArtifactChangeTarget, ArtifactService } from "./ArtifactService";
 
 /**
  * Owns the per-editor {@link ModelerSession} map that drives echo
@@ -34,20 +16,15 @@ import { ArtifactChangeTarget, ArtifactService } from "./ArtifactService";
  * extension writes back, so the resulting `onDidChangeTextDocument` event
  * is skipped by {@link display} instead of being re-rendered.
  */
-export class BpmnModelerService implements ArtifactChangeTarget {
+export class BpmnModelerService {
     private readonly sessions: Map<string, ModelerSession> = new Map();
 
     constructor(
         private readonly editorStore: EditorStore,
         private readonly vsDocument: VsCodeDocument,
-        private readonly vsSettings: VsCodeSettings,
-        private readonly notifier: VsCodeNotifier,
         private readonly picker: VsCodePicker,
-        private readonly clipboard: VsCodeClipboard,
-        private readonly artifactSvc: ArtifactService,
         private readonly statusBar: VsCodeStatusBar,
-        private readonly vsWorkspace: VsCodeWorkspace,
-        private readonly panelStateRepo: PropertiesPanelStateRepository,
+        private readonly notifier: VsCodeNotifier,
     ) {}
 
     registerSession(editorId: string): void {
@@ -144,153 +121,6 @@ export class BpmnModelerService implements ArtifactChangeTarget {
         }
     }
 
-    async setElementTemplates(editorId: string): Promise<boolean> {
-        this.statusBar.showElementTemplatesLoading();
-        try {
-            const documentDir = posix.dirname(this.vsDocument.getFilePath(editorId));
-
-            const [artifacts] = await this.artifactSvc.getArtifactPaths(documentDir);
-
-            const parsed = await Promise.all(
-                artifacts.map(async (a) => {
-                    try {
-                        return JSON.parse(await this.artifactSvc.readFile(a));
-                    } catch (error) {
-                        this.notifier.logError(
-                            new Error(
-                                `Failed to parse element template "${a}": ${(error as Error).message}`,
-                            ),
-                        );
-                        return [];
-                    }
-                }),
-            );
-            const sorted = parsed
-                .flat()
-                .sort((a: Record<string, unknown>, b: Record<string, unknown>) =>
-                    String(a.name ?? "").localeCompare(String(b.name ?? "")),
-                );
-
-            if (await this.editorStore.postMessage(editorId, new ElementTemplatesQuery(sorted))) {
-                this.statusBar.showElementTemplatesReady(sorted.length);
-                if (artifacts.length > 0) {
-                    this.notifier.logInfo(`${artifacts.length} element templates are set.`);
-                }
-                return true;
-            } else {
-                this.statusBar.hideElementTemplatesStatus();
-                return this.handleError(new Error("Setting the `elementTemplates` failed."));
-            }
-        } catch (error) {
-            this.statusBar.hideElementTemplatesStatus();
-            return this.handleError(error as Error);
-        }
-    }
-
-    async setSettings(editorId: string): Promise<boolean> {
-        try {
-            const settings = new SettingBuilder()
-                .alignToOrigin(this.vsSettings.getAlignToOrigin())
-                .showTransactionBoundaries(this.vsSettings.getShowTransactionBoundaries())
-                .colorTheme(this.vsSettings.getColorTheme())
-                .favouriteBpmnElements(this.vsSettings.getFavouriteBpmnElements())
-                .buildBpmnModeler();
-
-            if (
-                await this.editorStore.postMessage(
-                    editorId,
-                    new BpmnModelerSettingQuery({
-                        alignToOrigin: settings.alignToOrigin,
-                        showTransactionBoundaries: settings.showTransactionBoundaries,
-                        colorTheme: settings.colorTheme,
-                        favouriteBpmnElements: settings.favouriteBpmnElements,
-                    }),
-                )
-            ) {
-                return true;
-            } else {
-                return this.handleError(new Error("Unable to set preferences."));
-            }
-        } catch (error) {
-            this.notifier.logError(error as Error);
-            return false;
-        }
-    }
-
-    /**
-     * Sync read so the webview HTML can be pre-rendered with the correct
-     * collapsed state and the panel never flashes visible before
-     * {@link sendPropertiesPanelState} delivers the value over the channel.
-     */
-    getPersistedPanelVisibility(): boolean {
-        return this.panelStateRepo.getVisibility();
-    }
-
-    async sendPropertiesPanelState(editorId: string): Promise<boolean> {
-        try {
-            const visible = this.panelStateRepo.getVisibility();
-            return await this.editorStore.postMessage(
-                editorId,
-                new PropertiesPanelStateQuery(visible),
-            );
-        } catch (error) {
-            this.notifier.logError(error as Error);
-            return false;
-        }
-    }
-
-    /**
-     * Intentionally does not re-broadcast to other open webviews: each
-     * webview is authoritative over its own panel, so hiding it in one
-     * side-by-side editor must not close it in its neighbour.
-     */
-    async setPropertiesPanelVisibility(visible: boolean): Promise<void> {
-        try {
-            await this.panelStateRepo.setVisibility(visible);
-        } catch (error) {
-            this.notifier.logError(error as Error);
-        }
-    }
-
-    /**
-     * The extension host mediates clipboard access: VS Code sandboxed iframes
-     * lack `clipboard-read` / `clipboard-write` permissions.
-     */
-    async readClipboard(editorId: string): Promise<boolean> {
-        try {
-            const text = await this.clipboard.readClipboard();
-            return await this.editorStore.postMessage(editorId, new ClipboardQuery(text));
-        } catch (error) {
-            this.notifier.logError(error as Error);
-            return false;
-        }
-    }
-
-    async readTextClipboard(editorId: string): Promise<boolean> {
-        try {
-            const text = await this.clipboard.readClipboard();
-            return await this.editorStore.postMessage(editorId, new TextClipboardQuery(text));
-        } catch (error) {
-            this.notifier.logError(error as Error);
-            return false;
-        }
-    }
-
-    async writeClipboard(text: string): Promise<void> {
-        try {
-            await this.clipboard.writeClipboard(text);
-        } catch (error) {
-            this.notifier.logError(error as Error);
-        }
-    }
-
-    setLanguage(editorId: string): void {
-        const locale = this.vsSettings.getLanguage();
-        this.editorStore.postMessage(editorId, new LanguageQuery(locale)).catch((error) => {
-            this.notifier.logError(error instanceof Error ? error : new Error(String(error)));
-        });
-    }
-
     async changeEngineVersion(editorId: string): Promise<boolean> {
         try {
             const doc = new BpmnDocument(this.vsDocument.getContent(editorId));
@@ -310,160 +140,6 @@ export class BpmnModelerService implements ArtifactChangeTarget {
             }
             return this.handleError(error as Error);
         }
-    }
-
-    /**
-     * Same-engine only — no cross-platform migration (C7↔C8).
-     */
-    async migrateAllDiagrams(): Promise<boolean> {
-        try {
-            const paths = await this.vsWorkspace.findFiles("**/*.bpmn");
-            if (paths.length === 0) {
-                this.notifier.showInfo("No BPMN files found in the workspace.");
-                return false;
-            }
-
-            const plan = await this.buildMigrationPlan(paths);
-            if (plan.isEmpty()) {
-                this.notifier.showInfo(
-                    "Could not detect the engine for any BPMN file in the workspace.",
-                );
-                return false;
-            }
-
-            if (plan.undetected.length > 0) {
-                this.notifier.logWarning(
-                    `Skipped ${plan.undetected.length} file(s) with undetectable engine: ${plan.undetected.join(", ")}`,
-                );
-            }
-
-            let scope: MigrationScope;
-            if (plan.hasBothPlatforms()) {
-                scope = await this.picker.pickMigrationScope(
-                    plan.c7Files.length,
-                    plan.c8Files.length,
-                );
-            } else if (plan.hasC7()) {
-                scope = "c7";
-            } else {
-                scope = "c8";
-            }
-
-            // Collect all input before any writes: document-change listeners
-            // triggered by a write would steal focus and dismiss a subsequent
-            // QuickPick.
-            let c7Version: string | undefined;
-            let c8Version: string | undefined;
-
-            if (scope === "c7" || scope === "both") {
-                c7Version = await this.picker.pickEngineVersion("c7", getVersions("c7"));
-            }
-            if (scope === "c8" || scope === "both") {
-                c8Version = await this.picker.pickEngineVersion("c8", getVersions("c8"));
-            }
-
-            const summaryParts: string[] = [];
-
-            if (c7Version) {
-                const c7Updated = await this.applyVersionUpdate(plan.c7Files, c7Version, "c7");
-                if (c7Updated > 0) {
-                    summaryParts.push(`${c7Updated} diagram(s) to Camunda 7 (${c7Version})`);
-                }
-            }
-
-            if (c8Version) {
-                const c8Updated = await this.applyVersionUpdate(plan.c8Files, c8Version, "c8");
-                if (c8Updated > 0) {
-                    summaryParts.push(`${c8Updated} diagram(s) to Camunda 8 (${c8Version})`);
-                }
-            }
-
-            if (summaryParts.length > 0) {
-                this.notifier.showInfo(`Updated ${summaryParts.join(" and ")}.`);
-            } else {
-                this.notifier.showInfo("All diagrams are already at the selected version.");
-            }
-
-            return true;
-        } catch (error) {
-            if (error instanceof UserCancelledError) {
-                return false;
-            }
-            return this.handleError(error as Error);
-        }
-    }
-
-    private async buildMigrationPlan(paths: string[]): Promise<MigrationPlan> {
-        const c7Files: BpmnFileEntry[] = [];
-        const c8Files: BpmnFileEntry[] = [];
-        const undetected: string[] = [];
-
-        for (const filePath of paths) {
-            const content = await this.vsWorkspace.readFile(filePath);
-            const doc = new BpmnDocument(content);
-            try {
-                const platform = doc.detectPlatform();
-                const entry: BpmnFileEntry = { path: filePath, document: doc, platform };
-                if (platform === "c7") {
-                    c7Files.push(entry);
-                } else {
-                    c8Files.push(entry);
-                }
-            } catch {
-                undetected.push(filePath);
-            }
-        }
-
-        return new MigrationPlan(c7Files, c8Files, undetected);
-    }
-
-    /**
-     * Files open in an editor are updated via {@link VsCodeDocument.write};
-     * files only on disk are written via {@link VsCodeWorkspace.writeFile}.
-     */
-    private async applyVersionUpdate(
-        files: readonly BpmnFileEntry[],
-        targetVersion: string,
-        platform: Engine,
-    ): Promise<number> {
-        let updatedCount = 0;
-
-        for (const file of files) {
-            const currentVersion = file.document.detectPlatformVersion();
-            if (currentVersion === targetVersion) {
-                continue;
-            }
-
-            let updatedDoc: BpmnDocument;
-            if (currentVersion === undefined) {
-                const platformName = platform === "c7" ? "Camunda Platform" : "Camunda Cloud";
-                const schema =
-                    platform === "c7"
-                        ? `xmlns:camunda="http://camunda.org/schema/1.0/bpmn"`
-                        : `xmlns:zeebe="http://camunda.org/schema/zeebe/1.0"`;
-                updatedDoc = file.document.withExecutionPlatform(
-                    platformName,
-                    targetVersion,
-                    schema,
-                );
-                this.notifier.logWarning(
-                    `Added missing executionPlatform attribute to: ${file.path}`,
-                );
-            } else {
-                updatedDoc = file.document.withVersion(targetVersion);
-            }
-
-            const editorId = this.editorStore.findEditorIdByPath(file.path);
-            if (editorId !== undefined) {
-                await this.vsDocument.write(editorId, updatedDoc.xml);
-            } else {
-                await this.vsWorkspace.writeFile(file.path, updatedDoc.xml);
-            }
-
-            updatedCount++;
-        }
-
-        return updatedCount;
     }
 
     private handleError(error: Error): boolean {
